backend/requirement.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
kavenegar==1.1.2
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
